+++ bin/e2e-test-client/src/tests/contracts.rs
@@ -1,10 +1,5 @@
 use crate::test_context::TestContext;
 use fuel_core_chain_config::ContractConfig;
-use fuel_core_types::fuel_crypto::rand::{
-    rngs::StdRng,
-    Rng,
-    SeedableRng,
-};
 use libtest_mimic::Failed;
 use std::time::Duration;
 use tokio::time::timeout;
@@ -19,17 +14,18 @@ pub async fn deploy_large_contract(ctx: &TestContext) -> Result<(), Failed> {
     } else {
         vec![0u8; 1024 * 1024]
     };
-    let mut rng = StdRng::seed_from_u64(2222);
     let mut contract_config = ContractConfig {
         contract_id: Default::default(),
         code: bytecode,
-        salt: rng.gen(),
         ..Default::default()
     };
+    let salt = Default::default();
     let storage_slots = vec![];
-    contract_config.update_contract_id(&storage_slots);
+    contract_config.update_contract_id(salt, &storage_slots);
 
-    let deployment_request = ctx.bob.deploy_contract(contract_config, storage_slots);
+    let deployment_request =
+        ctx.bob
+            .deploy_contract(contract_config, salt, storage_slots);
 
     // wait for contract to deploy in 5 minutes, because 16mb takes a lot of time.
     timeout(Duration::from_secs(300), deployment_request).await??;