+++ crates/fuel-core/src/service/genesis.rs
@@ -17,7 +17,6 @@ use fuel_core_storage::{
     tables::{
         Coins,
         ConsensusParametersVersions,
-        ContractsInfo,
         ContractsLatestUtxo,
         ContractsRawCode,
         Messages,
@@ -52,10 +51,7 @@ use fuel_core_types::{
     },
     entities::{
         coins::coin::Coin,
-        contract::{
-            ContractUtxoInfo,
-            ContractsInfoType,
-        },
+        contract::ContractUtxoInfo,
         message::Message,
     },
     fuel_tx::Contract,
@@ -246,7 +242,6 @@ fn init_contract(
     height: BlockHeight,
 ) -> anyhow::Result<()> {
     let contract = Contract::from(contract_config.code.as_slice());
-    let salt = contract_config.salt;
     let contract_id = contract_config.contract_id;
     #[allow(clippy::cast_possible_truncation)]
     let utxo_id = contract_config.utxo_id();
@@ -267,14 +262,6 @@ fn init_contract(
         return Err(anyhow!("Contract code should not exist"));
     }
 
-    // insert contract salt
-    if transaction
-        .storage::<ContractsInfo>()
-        .insert(&contract_id, &ContractsInfoType::V1(salt.into()))?
-        .is_some()
-    {
-        return Err(anyhow!("Contract info should not exist"));
-    }
     if transaction
         .storage::<ContractsLatestUtxo>()
         .insert(
@@ -576,7 +563,6 @@ mod tests {
             contracts: vec![ContractConfig {
                 contract_id,
                 code: contract.into(),
-                salt,
                 tx_id: rng.gen(),
                 output_index: rng.gen(),
                 tx_pointer_block_height: 0.into(),
@@ -672,7 +658,6 @@ mod tests {
             contracts: vec![ContractConfig {
                 contract_id,
                 code: contract.into(),
-                salt,
                 ..Default::default()
             }],
             contract_balance: vec![contract_balance],
@@ -739,14 +724,12 @@ mod tests {
             asset_id: test_asset_id,
             amount: test_balance,
         }];
-        let salt: Salt = rng.gen();
         let contract = Contract::from(op::ret(0x10).to_bytes().to_vec());
 
         let state = StateConfig {
             contracts: vec![ContractConfig {
                 contract_id: ContractId::from(*contract_id),
                 code: contract.into(),
-                salt,
                 // set txpointer height > genesis height
                 tx_pointer_block_height: BlockHeight::from(11u32),
                 tx_pointer_tx_idx: 0,
@@ -793,7 +776,6 @@ mod tests {
         ContractConfig {
             contract_id,
             code: contract.into(),
-            salt,
             ..Default::default()
         }
     }